Fixed error running torch nn with "genetic" method
If you tried to generate a counterfactual with a torch neural network, built to solve a multiclass classification problem, with the "genetic" method you obtained an error due to a shapes mismatch. Changing the code, np.argmax will return only one value for each sample, as expected, and not a value (logit or probability) for each class.

Signed-off-by: Daniele Morotti <58258368+DanieleMorotti@users.noreply.github.com>
Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>