Implement CounterPlots to DiCE CF generation
This commit adds the CounterPlots package to the DiCE CF generation package. Basically, the `CounterfactualExplanations` object was changed to now include a new method `generate_counterplots` which takes factual, counterfactual, and DataFrame info to generate the counterfactual analysis. Also, this object now asks for a prediction function, therefore, the `explainer_base`, `dice_tensorflow1`, and `dice_tensorflow2` now include this function to create the object.