Fixed the counterfactual generation for pytorch nn when using 'genetic' method (#362)
* Fixed error running torch nn with "genetic" method

If you tried to generate a counterfactual with a torch neural network, built to solve a multiclass classification problem, with the "genetic" method you obtained an error due to a shapes mismatch. Changing the code, np.argmax will return only one value for each sample, as expected, and not a value (logit or probability) for each class.

Signed-off-by: Daniele Morotti <58258368+DanieleMorotti@users.noreply.github.com>
Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

* Fixed error for the cf creation with torch and genetic method.

When you used the 'genetic' method to create the counterfactual, if you're running a pytorch neural network for a multiclass classification problem, it raised an error due to a mismatch between shapes. It expected one value for each sample but it got a value for each class.

Signed-off-by: Daniele Morotti <58258368+DanieleMorotti@users.noreply.github.com>
Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

* Changed the correction using the appropriate function of the class

Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

* Added a test for the usage of pytorch nn with genetic algorithm

Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

* Fixed the unit test about multiclassification nn

Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

* Fixed isort error in the test

Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

* Fixed flake8 errors in test for multiclass nn

Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>

---------

Signed-off-by: Daniele Morotti <58258368+DanieleMorotti@users.noreply.github.com>
Signed-off-by: DanieleMorotti <daniele.morotti.99@gmail.com>
Co-authored-by: Gaurav Gupta <47334368+gaugup@users.noreply.github.com>